# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/payments/verify": {
      "post": {
        "operationId": "PaymentsController_verifyPayment",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/booking/success": {
      "get": {
        "operationId": "PaymentsController_bookingSuccess",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_handleWebhook",
        "parameters": [
          {
            "name": "x-paystack-signature",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/booking/details": {
      "get": {
        "operationId": "PaymentsController_getBookingDetails",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Payments"
        ]
      }
    },
...
    "/api/transactions": {
      "get": {
        "operationId": "TransactionsController_getTransactions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user transactions",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/summary": {
      "get": {
        "operationId": "TransactionsController_getTransactionSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Transaction summary retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction summary",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/{reference}": {
      "get": {
        "operationId": "TransactionsController_getTransactionByReference",
        "parameters": [
          {
            "name": "reference",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully"
          },
          "404": {
            "description": "Transaction not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transaction by reference",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/type/{type}": {
      "get": {
        "operationId": "TransactionsController_getTransactionsByType",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transactions by type",
        "tags": [
          "Transactions"
        ]
      }
    },
    "/api/transactions/status/{status}": {
      "get": {
        "operationId": "TransactionsController_getTransactionsByStatus",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get transactions by status",
        "tags": [
          "Transactions"
        ]
      }
    },
...
    "/api/admin/transactions/{id}": {
      "get": {
        "operationId": "AdminTransactionsController_getTransactionById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Transaction ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get specific transaction by ID",
        "tags": [
          "Admin Transactions"
        ]
      }
    },
...
  },
  "info": {
    "title": "CvestAPI",
    "description": "API documentation for CvestAPI",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}